+++ src/Microsoft.SqlTools.Serialization/Program.cs
@@ -3,9 +3,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using Microsoft.SqlTools.Credentials.Utility;
 using Microsoft.SqlTools.ServiceLayer.SqlContext;
 using Microsoft.SqlTools.Utility;
+using Microsoft.SqlTools.Serialization.Utility;
 
 namespace Microsoft.SqlTools.Serialization
 {
