+++ src/Microsoft.SqlTools.Serialization/Utility/CommandOptions.cs
@@ -1,11 +1,11 @@
-//
+﻿//
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
 using System;
 
-namespace Microsoft.SqlTools.Credentials.Utility
+namespace Microsoft.SqlTools.Serialization.Utility
 {
     /// <summary>
     /// The command-line options helper class.
@@ -57,7 +57,7 @@ namespace Microsoft.SqlTools.Credentials.Utility
             }
         }
 
-        internal string ErrorMessage { get; private set; }
+        public string ErrorMessage { get; private set; }
 
         /// <summary>
         /// Whether diagnostic logging is enabled
@@ -77,7 +77,7 @@ namespace Microsoft.SqlTools.Credentials.Utility
             get
             {
                 var str = string.Format("{0}" + Environment.NewLine +
-                    "Microsoft.SqlTools.Credentials " + Environment.NewLine +
+                    "Microsoft.SqlTools.Serialization " + Environment.NewLine +
                     "   Options:" + Environment.NewLine +
                     "        [--enable-logging]" + Environment.NewLine +
                     "        [--help]" + Environment.NewLine,
